app/requirements.txt
argh==0.26.2
asgiref==3.2.10
attrs==19.3.0
backoff==1.10.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-dramatiq==0.9.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
dramatiq==1.9.0
drf-yasg==1.17.1
feedparser==5.2.1
gevent==20.6.2
greenlet==0.4.16
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
openapi-codec==1.3.2
packaging==20.4
pathtools==0.1.2
pika==1.1.0
pluggy==0.13.1
prometheus-client==0.8.0
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
watchdog==0.8.3
watchdog-gevent==0.1.0
webencodings==0.5.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
